requirements.txt
python-telegram-bot==21.5
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.9.3
seaborn==0.13.2
reportlab==4.2.5
APScheduler==3.10.4
